LPS-171139 search-experiences-service: Rename package to align with the APIs
